requirements.txt
opencv-python==4.8.1.78
cvzone==1.6.1
Pillow==10.1.0
requests==2.31.0
mediapipe==0.10.8
numpy==1.24.3